README.md
This is the fourth course in the Data Science Specialization. Exploratory data analysis (EDA) is a key element of data science because it allows you to develop a rough idea of what your data look like and what kinds of questions might be answered by them. EDA is often the "fun part" of data analysis, where you get to play around with the data and, well, explore!